+++ packages/vercel-workflows/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @yolk-sdk/vercel-workflows
 
+## 0.1.0-canary.24
+
 ## 0.1.0-canary.23
 
 ## 0.1.0-canary.22